+++ api.go
@@ -14,6 +14,7 @@ import (
 
 	"github.com/dgraph-io/dgraph/v24/dql"
 	"github.com/dgraph-io/dgraph/v24/schema"
+	"github.com/hypermodeinc/modusdb/api/utils"
 )
 
 func Create[T any](db *DB, object T, ns ...uint64) (uint64, T, error) {
@@ -66,7 +67,7 @@ func Upsert[T any](db *DB, object T, ns ...uint64) (uint64, T, bool, error) {
 		return 0, object, false, err
 	}
 
-	gid, cf, err := getUniqueConstraint[T](object)
+	gid, cf, err := GetUniqueConstraint[T](object)
 	if err != nil {
 		return 0, object, false, err
 	}
@@ -85,13 +86,13 @@ func Upsert[T any](db *DB, object T, ns ...uint64) (uint64, T, bool, error) {
 
 	if gid != 0 {
 		gid, _, err = getByGidWithObject[T](ctx, n, gid, object)
-		if err != nil && err != ErrNoObjFound {
+		if err != nil && err != utils.ErrNoObjFound {
 			return 0, object, false, err
 		}
 		wasFound = err == nil
 	} else if cf != nil {
 		gid, _, err = getByConstrainedFieldWithObject[T](ctx, n, *cf, object)
-		if err != nil && err != ErrNoObjFound {
+		if err != nil && err != utils.ErrNoObjFound {
 			return 0, object, false, err
 		}
 		wasFound = err == nil
